fix: Address code-review findings on the #437 Dock-icon fix
## Summary
- A medium-effort code review of the #437 fix surfaced a real ordering race and two doc-accuracy issues; this addresses them before merge.

## Changes
- Restore an explicit `syncAgentActivationPolicy()` call after the library-restore logic in the display window's close handler, closing a race where the new global `willClose` observer's independently-scheduled reconcile Task had no guaranteed ordering against the restore, which could flip the Dock icon to `.accessory` and back within the same runloop pass.
- Correct doc comments (AppDelegate.swift, ARCHITECTURE.md) that overstated `isUntrackedUserPanel` covering the toolbar customization palette — AppKit presents that palette as a sheet attached to its target window, so it can never be the sole surviving window; the About panel is the actual motivating case.
- Fix a doc comment citing the wrong install site for `globalWindowCloseObserver` (`applicationDidFinishLaunching` vs. actual `startResidentApp`).
- Extract the new test's window-construction helper into a shared `makeTestWindow(styleMask:)` in `KernovaTests/TestHelpers.swift` instead of duplicating it inline, per the existing test-helper reuse convention.
- Drop the unnecessary `nonisolated` from `isUntrackedUserPanel`, whose only callers (production + test) are already `@MainActor`.

## Test plan
- [x] `make lint`
- [x] `make test` (full suite)
